backend/requirements.txt
asgiref==3.9.1
django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
ecdsa==0.19.1
numpy==2.3.1
pandas==2.3.1
pyasn1==0.6.1
python-dateutil==2.9.0.post0
python-jose==3.5.0
pytz==2025.2
rsa==4.9.1
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
